Fixed Typing for real

+++ main.py
@@ -3,7 +3,7 @@
 # This program will calculate a student's GPA from 3 classes
 
 def result2GP(letterGrade):
-  GP = 0
+  GP = 0.0
   if letterGrade[0] == 'A':
     GP += 4.0
   elif letterGrade[0] == 'B':
@@ -12,9 +12,9 @@ def result2GP(letterGrade):
     GP += 2.0
   elif letterGrade[0] == 'D':
     GP += 1.0
-  if letterGrade[1] == '+':
+  if letterGrade[1] == "+":
     GP += 0.33
-  elif letterGrade[1] == '-':
+  elif letterGrade[1] == "-":
     GP -= 0.33
   return GP
 
@@ -25,8 +25,8 @@ totalGP = 0
 totalWeight = 0
 
 while i != 3:
-  scores[i] = int(result2GP(input(f"Enter your course {i + 1} letter grade: ") + " "))
-  weights[i] = int(input(f"Enter your course {i + 1} credit: "))
+  scores[i] = float(result2GP(input(f"Enter your course {i + 1} letter grade: ") + " "))
+  weights[i] = float(input(f"Enter your course {i + 1} credit: "))
   print(f"Grade point for course {i+1} is: {scores[i]}")
   totalGP += scores[i]*weights[i]
   totalWeight += weights[i]
